+++ tests/test_data_models.py
@@ -89,7 +89,7 @@ class TestBenchmarkResult:
             max_sequence_length=100,
             vocabulary_size=100,
             merge=False,
-            dataset_languages=["da"],
+            languages=["da"],
             task="task",
             results=dict(),
         )
@@ -115,7 +115,7 @@ class TestBenchmarkResult:
         assert benchmark_result.max_sequence_length == 100
         assert benchmark_result.vocabulary_size == 100
         assert benchmark_result.merge is False
-        assert benchmark_result.dataset_languages == ["da"]
+        assert benchmark_result.languages == ["da"]
         assert benchmark_result.task == "task"
         assert benchmark_result.results == dict()
         assert benchmark_result.euroeval_version == __version__
@@ -132,7 +132,7 @@ class TestBenchmarkResult:
                     num_model_parameters=100,
                     max_sequence_length=100,
                     vocabulary_size=100,
-                    dataset_languages=["da"],
+                    languages=["da"],
                     task="task",
                     results=dict(),
                 ),
@@ -147,7 +147,7 @@ class TestBenchmarkResult:
                     num_model_parameters=100,
                     max_sequence_length=100,
                     vocabulary_size=100,
-                    dataset_languages=["da"],
+                    languages=["da"],
                     task="task",
                     results=dict(),
                 ),
@@ -160,7 +160,7 @@ class TestBenchmarkResult:
                     num_model_parameters=100,
                     max_sequence_length=100,
                     vocabulary_size=100,
-                    dataset_languages=["da"],
+                    languages=["da"],
                     task="task",
                     results=dict(),
                 ),
@@ -175,7 +175,7 @@ class TestBenchmarkResult:
                     num_model_parameters=100,
                     max_sequence_length=100,
                     vocabulary_size=100,
-                    dataset_languages=["da"],
+                    languages=["da"],
                     task="task",
                     results=dict(),
                 ),
@@ -188,7 +188,7 @@ class TestBenchmarkResult:
                     num_model_parameters=100,
                     max_sequence_length=100,
                     vocabulary_size=100,
-                    dataset_languages=["da"],
+                    languages=["da"],
                     task="task",
                     results=dict(),
                 ),
@@ -203,7 +203,7 @@ class TestBenchmarkResult:
                     num_model_parameters=100,
                     max_sequence_length=100,
                     vocabulary_size=100,
-                    dataset_languages=["da"],
+                    languages=["da"],
                     task="task",
                     results=dict(),
                 ),
@@ -215,7 +215,7 @@ class TestBenchmarkResult:
                     num_model_parameters=100,
                     max_sequence_length=100,
                     vocabulary_size=100,
-                    dataset_languages=["da"],
+                    languages=["da"],
                     task="task",
                     results=dict(),
                 ),
@@ -230,7 +230,7 @@ class TestBenchmarkResult:
                     num_model_parameters=100,
                     max_sequence_length=100,
                     vocabulary_size=100,
-                    dataset_languages=["da"],
+                    languages=["da"],
                     task="task",
                     results=dict(),
                 ),
@@ -259,7 +259,7 @@ class TestBenchmarkResult:
             dict(
                 dataset=benchmark_result.dataset,
                 task=benchmark_result.task,
-                dataset_languages=benchmark_result.dataset_languages,
+                languages=benchmark_result.languages,
                 model=benchmark_result.model,
                 results=benchmark_result.results,
                 num_model_parameters=benchmark_result.num_model_parameters,
@@ -295,7 +295,7 @@ class TestBenchmarkParametersAreConsistent:
         )
         benchmarker_init_params = set(
             inspect.signature(Benchmarker.__init__).parameters.keys()
-        ) - {"self"}
+        ) - {"self", "batch_size", "dataset_language", "model_language"}
         assert benchmark_config_params == benchmarker_init_params
 
     def test_config_params_is_the_same_as_benchmark_method(self) -> None:
@@ -305,7 +305,7 @@ class TestBenchmarkParametersAreConsistent:
         ) - {"run_with_cli"}
         benchmark_method_params = set(
             inspect.signature(Benchmarker.benchmark).parameters.keys()
-        ) - {"self", "model"}
+        ) - {"self", "model", "batch_size", "dataset_language", "model_language"}
         assert benchmark_config_params == benchmark_method_params
 
     def test_config_params_is_the_same_as_cli(
@@ -317,7 +317,9 @@ class TestBenchmarkParametersAreConsistent:
         ) - {"run_with_cli"}
         cli_benchmark_params = set(cli_params.keys()) - {
             "model",
-            "custom_datasets_file",
+            "batch_size",
+            "dataset_language",
+            "model_language",
             "help",
         }
         assert benchmark_config_params == cli_benchmark_params
@@ -326,8 +328,8 @@ class TestBenchmarkParametersAreConsistent:
         """Test that `BenchmarkConfigParams` agrees with `BenchmarkConfig`."""
         benchmark_config_params = set(
             inspect.signature(BenchmarkConfigParams).parameters.keys()
-        ) - {"dataset", "task", "language", "dataset_language", "model_language"}
+        ) - {"dataset", "task", "language"}
         benchmark_config_fields = set(
             inspect.signature(BenchmarkConfig).parameters.keys()
-        ) - {"datasets", "tasks", "dataset_languages", "model_languages"}
+        ) - {"datasets", "tasks", "languages"}
         assert benchmark_config_params == benchmark_config_fields